Initial commit for feature
Added context collection, which is an abstraction of base db context. all entries are key by a sessionid. All entires togehter, are a session context. SessionContext initializes late, so we have to make ISessionLoop, which should start if session is ready, make lazy. It must be lazy because the session details (like the exampels) arent set at the moment.
added sessioncontextmanager, which looks up via dbmanager if there are any context collections to know
Added example loop and project